+++ src/language/CompletionItemProvider.ts
@@ -4,6 +4,7 @@ import {
   CompletionItemProvider,
   Position,
   SnippetString,
+  SymbolKind,
   TextDocument,
 } from "vscode";
 import { isNumber, uniq } from "lodash";
@@ -15,6 +16,8 @@ import dirType from "./definitions/dirType";
 import nodeShapes from "./definitions/nodeShapes";
 import style from "./definitions/style";
 import SymbolProvider from "./SymbolProvider";
+import { DotSymbolDefinition } from "./SymbolDefinition";
+import getAttributeDetail from "./getAttributeDetail";
 
 export default class DotCompletionItemProvider implements CompletionItemProvider {
   private colors: CompletionItem[] = [];
@@ -44,13 +47,13 @@ export default class DotCompletionItemProvider implements CompletionItemProvider
   private specialAttributes: {[attribute: string] : string} = {};
 
   constructor() {
-    const names = {
+    /* const names = {
       G: "Root graph",
       N: "Nodes",
       E: "Edges",
       C: "Clusters",
       S: "Subgraphs",
-    };
+    }; */
 
     this.primitives = "node|edge|graph".split("|").map((type) => {
       const pack = new CompletionItem(type, CompletionItemKind.Constant);
@@ -90,6 +93,7 @@ export default class DotCompletionItemProvider implements CompletionItemProvider
     });
 
     attributeList.split("\n").forEach((al) => {
+      // eslint-disable-next-line no-unused-vars
       const [attribute, typeList, datatype, ...other] = al.split("|");
       const item = new CompletionItem(attribute, CompletionItemKind.Property);
       item.insertText = `${attribute}=`;
@@ -118,23 +122,36 @@ export default class DotCompletionItemProvider implements CompletionItemProvider
           this.specialAttributes[attribute] = datatype;
         }
       }
-      item.documentation = "Available on:";
+      item.documentation = getAttributeDetail(attribute) || "";
+
+      // item.documentation = "Available on:";
       for (let i = 0; i < typeList.length; i += 1) {
         (this.attributes as any)[typeList[i] as string].push(item);
-        item.documentation += `\n${(names as any)[typeList[i]]}`;
+        // item.documentation += `\n${(names as any)[typeList[i]]}`;
       }
     });
   }
 
   // eslint-disable-next-line class-methods-use-this
-  private provideSymbols(document: TextDocument): CompletionItem[] {
+  private provideSymbols(
+    document: TextDocument,
+    position: Position,
+  ): CompletionItem[] {
     const symProv = new SymbolProvider();
-    const symbols = symProv.provideSymbols(document);
+    const symbols = symProv.flatSymbols(symProv.provideSymbols(document));
 
     const suggestions: CompletionItem[] = [];
     const foundSymbols: string[] = [];
 
     symbols.forEach((symbol) => {
+      if (!((
+        ([DotSymbolDefinition.Node, DotSymbolDefinition.Graph] as unknown[] as SymbolKind[])
+      )).includes(symbol.kind)) {
+        return;
+      }
+      if (symbol.range.contains(position)) {
+        return;
+      }
       if (foundSymbols.includes(symbol.name)) {
         return;
       }
@@ -153,7 +170,7 @@ export default class DotCompletionItemProvider implements CompletionItemProvider
   ) : CompletionItem[] | undefined {
     const line = document.lineAt(position.line).text.substring(0, position.character);
 
-    let suggestions = this.provideSymbols(document);
+    let suggestions = this.provideSymbols(document, position);
 
     const reg = [
       {