+++ src/language/SymbolProvider.ts
@@ -1,4 +1,5 @@
 import {
+  DocumentSymbol,
   DocumentSymbolProvider,
   Location,
   Position,
@@ -6,11 +7,13 @@ import {
   Range,
   ReferenceProvider,
   RenameProvider,
-  SymbolInformation,
   SymbolKind,
   TextDocument,
   WorkspaceEdit,
 } from "vscode";
+
+import DotParser, { DocumentSymbolInformation } from "./DotParser";
+import { DotSymbolDefinition } from "./SymbolDefinition";
 
 export default class SymbolProvider
 implements
@@ -22,9 +25,16 @@ implements
     document: TextDocument,
     regex: RegExp,
     // eslint-disable-next-line no-unused-vars
-    func: (lineNumber: number, captures: RegExpExecArray, offset: number) => SymbolInformation[],
+    func: (
+      // eslint-disable-next-line no-unused-vars
+      lineNumber: number,
+      // eslint-disable-next-line no-unused-vars
+      captures: RegExpExecArray,
+      // eslint-disable-next-line no-unused-vars
+      offset: number
+      ) => DocumentSymbolInformation[],
   ) {
-    let symbols: SymbolInformation[] = [];
+    let symbols: DocumentSymbolInformation[] = [];
     for (let line = 0; line < document.lineCount; line += 1) {
       const t = document.lineAt(line).text;
       let match;
@@ -40,7 +50,7 @@ implements
 
   // find node definitions
   // /node\s*\[.*\]\s([\w\d]+|"(?:[^"\\]|\\.)*")/g
-  private findExplicitNodeDefinition(document: TextDocument): SymbolInformation[] {
+  private findExplicitNodeDefinition(document: TextDocument): DocumentSymbolInformation[] {
     const regex = /node\s*\[.*\]\s([\w\d]+|"(?:[^"\\]|\\.)*")/g;
 
     return this.lineIterator(
@@ -50,30 +60,36 @@ implements
         lineNumber: number,
         capture: RegExpExecArray,
         offset: number,
-      ): SymbolInformation[] => [new SymbolInformation(
-        capture[1],
-        SymbolKind.Variable,
-        "",
-        new Location(
-          document.uri,
-          new Range(
-            new Position(
-              lineNumber,
-              offset + capture[0].length - capture[1].length,
-            ),
-            new Position(
-              lineNumber,
-              offset + capture[0].length,
-            ),
+      ): DocumentSymbolInformation[] => {
+        const range = new Range(
+          new Position(
+            lineNumber,
+            offset + capture[0].length - capture[1].length,
           ),
-        ),
-      )],
+          new Position(
+            lineNumber,
+            offset + capture[0].length,
+          ),
+        );
+        const location = new Location(document.uri, range);
+
+        const ds = new DocumentSymbol(
+          capture[1],
+          "node",
+          SymbolKind.Variable,
+          range,
+          range,
+        );
+        (ds as DocumentSymbolInformation).location = location;
+        (ds as DocumentSymbolInformation).containerName = "";
+        return [ds as DocumentSymbolInformation];
+      },
     );
   }
 
   // find inline nodes
   // /^\s*((([\w\d]+)\s*)+)(?!.*(=|->|{|\[))/g
-  private findNodeDefinition(document: TextDocument): SymbolInformation[] {
+  private findNodeDefinition(document: TextDocument): DocumentSymbolInformation[] {
     const regex = /^\s*((([\w\d]+|"(?:[^"\\]|\\.)*")\s*)+)(?!.*(=|->|{|\[))/g;
     const regexSymbol = /([\w\d]+|"(?:[^"\\]|\\.)*")/g;
 
@@ -84,25 +100,30 @@ implements
         lineNumber: number,
         captures: RegExpExecArray,
         offset: number,
-      ): SymbolInformation[] => {
-        const symbols: SymbolInformation[] = [];
+      ): DocumentSymbolInformation[] => {
+        const symbols: DocumentSymbolInformation[] = [];
         let symbol;
         // eslint-disable-next-line no-cond-assign
         while ((symbol = regexSymbol.exec(captures[0])) != null) {
-          symbols.push(
-            new SymbolInformation(
-              symbol[0],
-              SymbolKind.Variable,
-              "",
-              new Location(
-                document.uri,
-                new Range(
-                  new Position(lineNumber, offset + symbol.index - 1),
-                  new Position(lineNumber, offset + symbol.index - 1 + symbol[0].length),
-                ),
-              ),
-            ),
+          const range = new Range(
+            new Position(lineNumber, offset + symbol.index - 1),
+            new Position(lineNumber, offset + symbol.index - 1 + symbol[0].length),
           );
+          const location = new Location(
+            document.uri,
+            range,
+          );
+          const ds = new DocumentSymbol(
+            symbol[0],
+            "node",
+            SymbolKind.Variable,
+            range,
+            range,
+          );
+          (ds as DocumentSymbolInformation).location = location;
+          (ds as DocumentSymbolInformation).containerName = "";
+          // return [ds as DocumentSymbolInformation];
+          symbols.push(ds as DocumentSymbolInformation);
         }
         return symbols;
       },
@@ -110,7 +131,7 @@ implements
   }
 
   // find Nodes with config
-  private findNodeDefinitionWithConfig(document: TextDocument) : SymbolInformation[] {
+  private findNodeDefinitionWithConfig(document: TextDocument) : DocumentSymbolInformation[] {
     const regex = /(?!\s*node)^\s*(([\w\d]+|"(?:[^"\\]|\\.)*"))\s*\[/g;
 
     return this.lineIterator(
@@ -120,30 +141,40 @@ implements
         lineNumber: number,
         capture: RegExpExecArray,
         offset: number,
-      ): SymbolInformation[] => [new SymbolInformation(
-        capture[1],
-        SymbolKind.Variable,
-        "",
-        new Location(
-          document.uri,
-          new Range(
-            new Position(
-              lineNumber,
-              offset + capture.index + (capture[0].length - capture[0].trim().length) - 1,
-            ),
-            new Position(
-              lineNumber,
-              offset + capture.index
-              + (capture[0].length - capture[0].trim().length)
-              + capture[1].length - 1,
-            ),
+      ): DocumentSymbolInformation[] => {
+        const range = new Range(
+          new Position(
+            lineNumber,
+            offset + capture.index + (capture[0].length - capture[0].trim().length) - 1,
           ),
-        ),
-      )],
+          new Position(
+            lineNumber,
+            offset + capture.index
+          + (capture[0].length - capture[0].trim().length)
+          + capture[1].length - 1,
+          ),
+        );
+        const location = new Location(
+          document.uri,
+          range,
+        );
+
+        const ds = new DocumentSymbol(
+          capture[1],
+          "node",
+          SymbolKind.Variable,
+          range,
+          range,
+        );
+        (ds as DocumentSymbolInformation).location = location;
+        (ds as DocumentSymbolInformation).containerName = "";
+
+        return [ds as DocumentSymbolInformation];
+      },
     );
   }
 
-  private findRegularSymbols(document: TextDocument) : SymbolInformation[] {
+  private findRegularSymbols(document: TextDocument) : DocumentSymbolInformation[] {
     const regex = /([\w\d]+|"(?:[^"\\]|\\.)*")(\s*(->|<-|--)\s*([\w\d]+|"(?:[^"\\]|\\.)*"))+/g;
     const regexSymbol = /[\w\d]+|"(?:[^"\\]|\\.)*"/g;
 
@@ -154,58 +185,98 @@ implements
         lineNumber: number,
         captures: RegExpExecArray,
         offset: number,
-      ) : SymbolInformation[] => {
-        const symbols: SymbolInformation[] = [];
+      ) : DocumentSymbolInformation[] => {
+        const symbols: DocumentSymbolInformation[] = [];
         let symbol;
         // eslint-disable-next-line no-cond-assign
         while ((symbol = regexSymbol.exec(captures[0])) != null) {
-          symbols.push(
-            new SymbolInformation(
-              symbol[0],
-              SymbolKind.Variable,
-              "",
-              new Location(
-                document.uri,
-                new Range(
-                  new Position(lineNumber, offset + symbol.index - 1),
-                  new Position(lineNumber, offset + symbol.index - 1 + symbol[0].length),
-                ),
-              ),
-            ),
+          const range = new Range(
+            new Position(lineNumber, offset + symbol.index - 1),
+            new Position(lineNumber, offset + symbol.index - 1 + symbol[0].length),
           );
+          const location = new Location(
+            document.uri,
+            range,
+          );
+          const ds = new DocumentSymbol(
+            symbol[0],
+            "node",
+            SymbolKind.Variable,
+            range,
+            range,
+          );
+          (ds as DocumentSymbolInformation).location = location;
+          (ds as DocumentSymbolInformation).containerName = "";
+          symbols.push(ds as DocumentSymbolInformation);
         }
         return symbols;
       },
     );
   }
 
-  public provideSymbols(document: TextDocument) : SymbolInformation[] {
-    let symbols : SymbolInformation[] = [];
+  // eslint-disable-next-line class-methods-use-this
+  public provideSymbols(document: TextDocument) : DocumentSymbolInformation[] {
+    let symbols : DocumentSymbolInformation[] = [];
 
-    symbols = symbols.concat(this.findExplicitNodeDefinition(document));
-    symbols = symbols.concat(this.findNodeDefinition(document));
-    symbols = symbols.concat(this.findRegularSymbols(document));
-    symbols = symbols.concat(this.findNodeDefinitionWithConfig(document));
+    try {
+      const dotParser = new DotParser();
+      dotParser.parse(document);
+      symbols = [(dotParser.getVscodeTypedAst() as DocumentSymbolInformation)];
+    } catch (e: any) {
+      symbols = symbols.concat(this.findExplicitNodeDefinition(document));
+      symbols = symbols.concat(this.findNodeDefinition(document));
+      symbols = symbols.concat(this.findRegularSymbols(document));
+      symbols = symbols.concat(this.findNodeDefinitionWithConfig(document));
+    }
 
     return symbols;
+  }
+
+  // eslint-disable-next-line class-methods-use-this
+  public flatSymbols(symbols: DocumentSymbolInformation[]) : DocumentSymbolInformation[] {
+    const res : DocumentSymbolInformation[] = [];
+
+    const crawler = (syms: DocumentSymbolInformation[]) => {
+      syms.forEach((symbol) => {
+        res.push(symbol);
+        if (symbol.children && symbol.children.length > 0) {
+          crawler(symbol.children);
+        }
+      });
+    };
+    crawler(symbols);
+
+    return res;
   }
 
   public provideDocumentSymbols(
     document: TextDocument,
     // token: CancellationToken,
-  ): Promise<SymbolInformation[]> {
+  ): Promise<DocumentSymbolInformation[]> {
     return Promise.resolve(this.provideSymbols(document));
   }
 
   // eslint-disable-next-line class-methods-use-this
-  private containingSymbol(symbols: SymbolInformation[], position: Position) {
-    let renameSymbol: SymbolInformation|undefined;
-    symbols.forEach((symbol) => {
-      if (symbol.location.range.contains(position)) {
-        renameSymbol = symbol;
-      }
-    });
-    return renameSymbol;
+  private containingSymbol(
+    symbols: DocumentSymbolInformation[],
+    position: Position,
+  ) {
+    let closestSymbol: DocumentSymbolInformation|undefined;
+
+    const crawler = (syms: DocumentSymbolInformation[]) => {
+      syms.forEach((symbol) => {
+        if (symbol.location.range.contains(position)
+        && (!closestSymbol || closestSymbol.range.contains(symbol.range))) {
+          closestSymbol = symbol;
+        }
+        if (symbol.children && symbol.children.length > 0) {
+          crawler(symbol.children);
+        }
+      });
+    };
+
+    crawler(symbols);
+    return closestSymbol;
   }
 
   public prepareRename(
@@ -215,7 +286,10 @@ implements
   ): ProviderResult<Range | { range: Range; placeholder: string }> {
     const symbols = this.provideSymbols(document);
     const renameSymbol = this.containingSymbol(symbols, position);
-    if (!renameSymbol) {
+    if (!renameSymbol
+      || !((
+        ([DotSymbolDefinition.Node, DotSymbolDefinition.Graph] as unknown[] as SymbolKind[])
+      )).includes(renameSymbol.kind)) {
       return Promise.reject(new Error("This can not be renamed."));
     }
 
@@ -232,30 +306,16 @@ implements
   // token: CancellationToken,
   ):
     Promise<WorkspaceEdit> {
-    return this.provideDocumentSymbols(document).then((symbols) => {
+    return this.provideReferences(document, position).then((locations) => {
       const edit:WorkspaceEdit = new WorkspaceEdit();
-
-      const renameSymbol = this.containingSymbol(symbols, position);
-      if (!renameSymbol) {
-        return Promise.reject();
-      }
 
       let newSymbolName = newName;
       if (!newName.match(/^[\w\d]+$/) && !(newName[0] === "\"" && newName[newName.length - 1] === "\"")) {
         newSymbolName = `"${newName.replace(/\\{0,1}"/g, "\\\"")}"`;
       }
 
-      symbols.forEach((symbol) => {
-        if (renameSymbol?.name === symbol.name) {
-          edit.replace(
-            document.uri,
-            symbol.location.range,
-            newSymbolName,
-          );
-        }
-      });
-
-      return Promise.resolve(edit);
+      locations.forEach((l) => edit.replace(document.uri, l.range, newSymbolName));
+      return edit;
     });
   }
 
@@ -267,18 +327,20 @@ implements
     // token: CancellationToken,
   ):
     Promise<Location[]> {
-    return this.provideDocumentSymbols(document).then((symbols) => {
-      const findSymbol = this.containingSymbol(symbols, position);
-      if (!findSymbol) {
-        return Promise.reject();
+    const symbols = this.provideSymbols(document);
+    const findSymbol = this.containingSymbol(symbols, position);
+    if (!findSymbol || !((
+      ([DotSymbolDefinition.Node, DotSymbolDefinition.Graph] as unknown[] as SymbolKind[])
+    )).includes(findSymbol.kind)) {
+      return Promise.reject();
+    }
+    const locations : Location[] = [];
+    const flatSymbols = this.flatSymbols(symbols);
+    flatSymbols.forEach((symbol) => {
+      if (findSymbol.name === symbol.name) {
+        locations.push(symbol.location);
       }
-      const locations : Location[] = [];
-      symbols.forEach((symbol) => {
-        if (findSymbol?.name === symbol.name) {
-          locations.push(symbol.location);
-        }
-      });
-      return Promise.resolve(locations);
     });
+    return Promise.resolve(locations);
   }
 }